+++ src/main/java/shop/mtcoding/bank/dto/user/UserRequest.java
@@ -1,0 +1,31 @@
+package shop.mtcoding.bank.dto.user;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import shop.mtcoding.bank.domain.user.User;
+import shop.mtcoding.bank.domain.user.UserEnum;
+
+public class UserRequest {
+
+    @Getter
+    @Setter
+    public static class JoinRequest {
+        // 유효성검사
+        private String username;
+        private String password;
+        private String email;
+        private String fullname;
+
+        public User toEntity(PasswordEncoder passwordEncoder) {
+            return User.builder()
+                    .username(username)
+                    .password(passwordEncoder.encode(password))
+                    .email(email)
+                    .fullname(fullname)
+                    .role(UserEnum.CUSTOMER)
+                    .build();
+        }
+    }
+
+}
